+++ src/MEGASync/transfers/gui/TransferManager.cpp
@@ -620,7 +620,7 @@ void TransferManager::refreshStateStats()
         else if(failedNumber != 0 || !MegaSyncApp->getStalledIssuesModel()->isEmpty())
         {
             leftFooterWidget = mUi->pSomeIssues;
-            mUi->bSomeIssues->setText(tr("Issue found", "", failedNumber));
+            mUi->bSomeIssues->setText(tr("Issue found", "", static_cast<int>(failedNumber)));
         }
         else if(processedNumber != 0)
         {
@@ -857,14 +857,14 @@ void TransferManager::refreshSpeed()
     mUi->wUpSpeed->setVisible(mTransfersCount.pendingUploads);
     if(mTransfersCount.pendingUploads)
     {
-        auto upSpeed (static_cast<unsigned long long>(mMegaApi->getCurrentUploadSpeed()));
+        auto upSpeed (mMegaApi->getCurrentUploadSpeed());
         mUi->lUpSpeed->setText(Utilities::getSizeString(upSpeed) + QLatin1String("/s"));
     }
 
     mUi->wDownSpeed->setVisible(mTransfersCount.pendingDownloads);
     if(mTransfersCount.pendingDownloads)
     {
-        auto dlSpeed (static_cast<unsigned long long>(mMegaApi->getCurrentDownloadSpeed()));
+        auto dlSpeed (mMegaApi->getCurrentDownloadSpeed());
         mUi->lDownSpeed->setText(Utilities::getSizeString(dlSpeed) + QLatin1String("/s"));
     }
 }
@@ -1301,9 +1301,8 @@ void TransferManager::checkActionAndMediaVisibility()
     }
 
     // Hide Media groupbox if no transfers (active or finished)
-    if (mUi->wTransfers->getCurrentTab() >= TransfersWidget::TYPES_TAB_BASE
-            || ((allTransfers +
-                completedTransfers + failedTransfers) > 0))
+    if (mUi->wTransfers->getCurrentTab() >= TransfersWidget::TYPES_TAB_BASE ||
+        ((allTransfers + completedTransfers + failedTransfers) > 0))
     {
         mUi->wMediaType->show();
     }
